+++ CHANGELOG.md
@@ -11,6 +11,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - SVG: `use_aux_axis_as_origin` option (#681)
 - Report: thickness units (#685)
 - PCB Print: Note about `colored_pads` and `colored_vias` side effects (#682)
+- Worksheet: undocumented font face and color now can be used, but not
+  correctly rendered (See #695)
 
 ### Fixed
 - PCB Print: allow specifying `repeat_for_layer` with empty `repeat_layers`.